internalize a bunch of utils
